0.3.2
Added in an "Empties" hierarchy object. It contains all of the necessary
empty objects for proper Scene organization ("Crates", "Score Zones",
"Walls", and "Floors"). This will slightly reduce the time it takes to
create levels AND minimize the risk of positioning errors.
-The crate rounding effect timer is now reset if the crate moves out of
a score zone (ChangeScore.cs).
-Added in 2-2
-Added in a GameManager object, along with scripts
-CamTricks.cs now works with the new GameManager to zoom out the game if
the previous level was zoomed in and zoom in the if the previous was
zoomed out
-